fix(core): resolve compile errors in religion/animation rework
- TechTreeDB: fully qualify TheWaningBorder.Data.BuildingCosts
- UnitAnimationSync: compare MinerState.State directly with MinerWorkState enum (not byte cast)
- SectPassiveHUD: add missing using TheWaningBorder.UI.Common for UIHelpers
- ActiveAbilityBar: add missing using TheWaningBorder.UI.Common for UIHelpers

+++ Assets/Scripts/Presentation/UnitAnimationSync.cs
@@ -144,7 +144,7 @@ namespace TheWaningBorder.Presentation
                 if (_em.HasComponent<MinerState>(LinkedEntity))
                 {
                     var miner = _em.GetComponentData<MinerState>(LinkedEntity);
-                    isWorking = miner.State == (byte)MinerWorkState.Gathering;
+                    isWorking = miner.State == MinerWorkState.Gathering;
                 }
 
                 // Builder: check active BuildOrder
